+++ src/OpenTelemetry.Api/CHANGELOG.md
@@ -7,8 +7,10 @@
 * Removed `IsOk` property from `Status` and fixed `StatusCode` enum values
   ([#1414](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1414))
 * `B3Propagator` now supports the value `true` to be passed in for the header
- `X-B3-Sampled`.
- ([#1413](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1413))
+  `X-B3-Sampled`.
+  ([#1413](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1413))
+* Moving grpc status and helper to grpc project
+  ([#1422](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1422))
 * Renamed TextMapPropagator to TraceContextPropagator, CompositePropapagor
   to CompositeTextMapPropagator. IPropagator is renamed to TextMapPropagator
   and changed from interface to abstract class.
@@ -92,7 +94,8 @@ Released 2020-08-28
   ([#994](https://github.com/open-telemetry/opentelemetry-dotnet/pull/994))
 * Changed `StartSpan` to not set the created span as Active to match the spec
   ([#994](https://github.com/open-telemetry/opentelemetry-dotnet/pull/994))
-* Updated System.Diagnostics.DiagnosticSource to version 5.0.0-preview.8.20407.11.
+* Updated System.Diagnostics.DiagnosticSource to version
+  5.0.0-preview.8.20407.11.
 * Removed `CorrelationContext` and added `Baggage`, an implementation of the
   [`Baggage
   API`](https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/baggage/api.md)
