+++ src/OpenTelemetry.Api/Trace/SpanHelper.cs
@@ -38,27 +38,5 @@ namespace OpenTelemetry.Trace
 
             return status;
         }
-
-        /// <summary>
-        /// Helper method that populates span properties from RPC status code according
-        /// to https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/trace/semantic_conventions/rpc.md#status.
-        /// </summary>
-        /// <param name="statusCode">RPC status code.</param>
-        /// <returns>Resolved span <see cref="Status"/> for the Grpc status code.</returns>
-        public static Status ResolveSpanStatusForGrpcStatusCode(int statusCode)
-        {
-            var status = Status.Error;
-
-            if (typeof(StatusCanonicalCode).IsEnumDefined(statusCode))
-            {
-                status = ((StatusCanonicalCode)statusCode) switch
-                {
-                    StatusCanonicalCode.Ok => Status.Unset,
-                    _ => Status.Error,
-                };
-            }
-
-            return status;
-        }
     }
 }
