+++ src/OpenTelemetry.Instrumentation.GrpcNetClient/GrpcTagHelper.cs
@@ -13,6 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // </copyright>
+
 using System.Diagnostics;
 using System.Text.RegularExpressions;
 using OpenTelemetry.Trace;
@@ -42,7 +43,7 @@ namespace OpenTelemetry.Instrumentation.GrpcNetClient
             var grpcStatusCodeTag = activity.GetTagValue(GrpcStatusCodeTagName);
             if (int.TryParse(grpcStatusCodeTag as string, out var statusCode))
             {
-                status = SpanHelper.ResolveSpanStatusForGrpcStatusCode(statusCode);
+                status = ResolveSpanStatusForGrpcStatusCode(statusCode);
             }
 
             return status;
@@ -63,6 +64,28 @@ namespace OpenTelemetry.Instrumentation.GrpcNetClient
                 rpcMethod = string.Empty;
                 return false;
             }
+        }
+
+        /// <summary>
+        /// Helper method that populates span properties from RPC status code according
+        /// to https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/trace/semantic_conventions/rpc.md#status.
+        /// </summary>
+        /// <param name="statusCode">RPC status code.</param>
+        /// <returns>Resolved span <see cref="Status"/> for the Grpc status code.</returns>
+        public static Status ResolveSpanStatusForGrpcStatusCode(int statusCode)
+        {
+            var status = Status.Error;
+
+            if (typeof(StatusCanonicalCode).IsEnumDefined(statusCode))
+            {
+                status = ((StatusCanonicalCode)statusCode) switch
+                {
+                    StatusCanonicalCode.Ok => Status.Unset,
+                    _ => Status.Error,
+                };
+            }
+
+            return status;
         }
     }
 }
